Lots of platform-specific fixes, mostly for OSX and Windows.

+++ src/treecore_bin_builder.cpp
@@ -329,7 +329,7 @@ int main( int argc, char const** argv )
         "#endif\r\n";
 
     if (!quiet && !dry_run)
-        Logger::writeToLog( "Total size of binary data: " + String( totalBytes ) + " bytes" );
+        Logger::writeToLog( "Total size of binary data: " + String( int64(totalBytes) ) + " bytes" );
 
     // clear stuffs to avoid some messages
     StringPool::releaseInstance();
